Make documentation platform open
Docusaurus is good but Read the Docs that is already used with API docs has benefits as well.
The decision needs to be done in collaboration with the core development team that will use
the selected platform in their daily work.

Also removed the goal of moving examples from the UG to Guides. Separate Guides are
awesome, but the UG needs examples as well!